refactor sd endpoint schemas and copy
